get number of rows and columns, filling matrix

+++ main.cpp
@@ -13,5 +13,19 @@ int main()
   Matrix mtx = Matrix(M, N);
   mtx.read();
   mtx.write();
+  std::cout << "Number of rows: " << mtx.getRows() << "\n";
+  std::cout << "Number of columns: " << mtx.getColumns() << "\n";
+  int a = 0;
+  std::cin >> a;
+  if (!std::cin)
+  {
+    std::cerr << "Wrong input!\n";
+    return 1;
+  }
+  else
+  {
+    mtx.filling(a);
+    mtx.write();
+  }
   mtx.destroy();
 }
